Refactor dj fetching for event list
