+++ webapp/src/contexts/useEventDjCache/eventDjCacheContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from "react"
-import { Dj, Event } from '../util/types';
+import { Dj, Event } from '../../util/types';
 import { DjCache, EventCache } from "./types";
 
 type EventDjCacheContextType = {
@@ -11,6 +11,7 @@ type EventDjCacheContextType = {
       djs: (Dj | "PENDING")[];
   } | null;
   getEventsByDjId: (djId: string) => Event[]; // Added function type
+  getPlayedDjsForEvent: (event: Event) => Dj[];
 };
 
 export const EventDjPlayMapperContext = createContext<EventDjCacheContextType | undefined>(undefined);
