+++ webapp/src/contexts/useEventDjCache/types.ts
@@ -1,4 +1,4 @@
-import { Dj, Event } from "../util/types";
+import { Dj, Event } from "../../util/types";
 
 export type DjCache = Map<string, Dj>;
 export type EventCache = Map<string, Event>;
